requirements/base.txt
Django==1.8.3
django-celery==3.1.16
django-activity-stream==0.6.0
django-auth-ldap3==0.9.2
gitpython
django_ace
djangorestframework
django-simple-history
reportlab
